requirements.txt
Flask==1.1.2
pymongo[tls,srv]==3.7.2
pytest==3.3.0
Flask_Bcrypt==0.7.1
python-dotenv==0.14.0
